refactor: split internal/ziplab/processor.go by concern
571 lines mixing pipeline orchestration with the zip mechanics it drives,
split along the seam the backlog identified:

  processor.go             218  the Step* pipeline stages + external command
  processor_zip.go         211  integrity test, extract, archive rewrite
  processor_adremoval.go   130  ad-file matching and removal
  processor_diz.go          49  FILE_ID.DIZ lookup

Pure move: no declaration body changed. Verified two ways — a go/parser
pairwise diff of every top-level declaration (20 old, 20 new, 0 missing,
0 mismatched, 0 extra), and a multiset diff of every comment line, which
shows nothing removed and only the four new per-file headers added.
No test file touched.
